requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document parsing
pypdf==3.17.1
python-docx==1.1.0
markdown==3.5.1
unstructured==0.11.2
pytesseract==0.3.10  # For OCR if needed
pdf2image==1.16.3

# Vector DB and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# LLM integration
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1  # For OpenAI-compatible APIs

# YandexGPT/GigaChat specific
requests==2.31.0
aiohttp==3.9.1

# Web interface
streamlit==1.29.0
streamlit-chat==0.1.1

# Utilities
tqdm==4.66.1
numpy==1.26.2
pandas==2.1.4
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.7.1
